refactor: move millennium tabs to src/tabs
